Unify VIC-20 and C64 keyboard stuff
